Modif bot discord

+++ chatBotProjet/backend/discordConfigurationBot/index.js
@@ -14,7 +14,7 @@ function botNotReady(err) {
   console.log('An error has occurred.', err);
 }
 const Discord = require('discord.js');
-const {prefix, token} = require('./config.json');
+const { prefix, token } = require('./config.json');
 const client = new Discord.Client();
 
 client.once('ready', () => {
@@ -25,22 +25,44 @@ client.once('ready', () => {
   console.log(`argv._[0]`, argv._[0])
 });
 
+function loadBrains() {
+  bot.loadFile('./brain2.rive').then(botReady).catch(botNotReady);
+}
 
 
 client.on('message', message => {
   if (!message.content.startsWith(prefix) || message.author.bot) return;
-  const args = message.content.slice(prefix.length).split(/ +/);
-  const command = args.shift().toLowerCase();
+  //for dialogue
+  const command = message.content.slice(prefix.length)
+  //for special command
+  const args = message.content.slice(prefix.length).trim().split(' ');
+  const specialCommandToken = args.shift().toLowerCase();
+  const specialCommand =  args.shift();
   try {
-    bot.reply('local - user', command).then(function(reply) {
-      console.log(`reply`, reply)
-      message.reply(`**${argv._[0]}** : ${reply}`);
-      if (command === 'disconnect') {
-        message.channel.send('Good bye !').then(m => {
-          client.destroy();
-        })
+    //si commande spéciale
+    if (specialCommandToken === 'sc') {
+      message.channel.send(`**Special command** called:\n**Command**: ${specialCommand}\n**Arguments**: ${args}`);
+      switch (specialCommand)
+      {
+        case 'load' :
+          loadBrains()
+          break;
+        default :
+          message.channel.send('Special command not recognize')
       }
-    });
+
+    }
+
+    //sinon dialogue
+    else {
+      bot.reply('local - user', command).then(function (reply) {
+        console.log(`reply`, reply)
+        message.reply(`**${argv._[0]}** : ${reply}`);
+        if (command === 'disconnect') {
+          client.destroy();
+        }
+      });
+    }
   } catch (error) {
     console.error(error);
   }
